pyproject.toml
[tool.poetry]
name = "hotels-application"
version = "0.1.0"
description = ""
authors = ["ol.n.nikiforova <ol.n.nikiforova@tinkoff.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.100.0"
uvicorn = "^0.28.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
email-validator = "^2.1.1"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"


[tool.poetry.group.format.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
